# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 建設業向け原価管理クラウドシステム API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cost-items:
    get:
      operationId: listCostItems
      tags: [cost-items]
      summary: 原価項目一覧取得
      parameters:
        - in: query
          name: projectId
          required: true
          schema:
            type: integer
        - in: query
          name: category
          schema:
            type: string
            enum: [material, labor, subcontract, expense]
      responses:
        "200":
          description: 原価項目一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostItemListResponse"
    post:
      operationId: createCostItem
      tags: [cost-items]
      summary: 原価項目登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCostItemRequest"
      responses:
        "201":
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostItem"
  /cost-items/{id}:
    put:
      operationId: updateCostItem
      tags: [cost-items]
      summary: 原価項目更新
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCostItemRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostItem"
    delete:
      operationId: deleteCostItem
      tags: [cost-items]
      summary: 原価項目削除
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 削除成功
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: 予算一覧取得
      parameters:
        - in: query
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 予算一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetListResponse"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: 予算登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetRequest"
      responses:
        "201":
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    put:
      operationId: updateBudget
      tags: [budgets]
      summary: 予算更新
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
...
  /dashboard/cost-by-category:
    get:
      operationId: getCostByCategory
      tags: [dashboard]
      summary: カテゴリ別原価集計
      parameters:
        - in: query
          name: projectId
          schema:
            type: integer
      responses:
        "200":
          description: カテゴリ別原価
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostByCategoryResponse"
  /dashboard/monthly-costs:
    get:
      operationId: getMonthlyCosts
      tags: [dashboard]
      summary: 月別原価推移取得
      parameters:
        - in: query
          name: projectId
          schema:
            type: integer
        - in: query
          name: year
          schema:
            type: integer
      responses:
        "200":
          description: 月別原価推移
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyCostsResponse"
  /dashboard/budget-vs-actual:
    get:
      operationId: getBudgetVsActual
      tags: [dashboard]
      summary: 予算実績対比取得
      parameters:
        - in: query
          name: projectId
          schema:
            type: integer
      responses:
        "200":
          description: 予算実績対比
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetVsActualResponse"
components:
  schemas:
...
    CostItem:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        category:
          type: string
          enum: [material, labor, subcontract, expense]
          description: 原価区分（材料費・労務費・外注費・経費）
        description:
          type: string
          description: 摘要
        vendor:
          type: string
          nullable: true
          description: 取引先
        quantity:
          type: number
          nullable: true
          description: 数量
        unit:
          type: string
          nullable: true
          description: 単位
        unitPrice:
          type: number
          nullable: true
          description: 単価
        amount:
          type: number
          description: 金額
        incurredDate:
          type: string
          format: date
          description: 発生日
        invoiceNumber:
          type: string
          nullable: true
          description: 伝票番号
        notes:
          type: string
          nullable: true
          description: 備考
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - category
        - description
        - amount
        - incurredDate
        - createdAt
        - updatedAt
    CostItemListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CostItem"
        total:
          type: integer
        totalAmount:
          type: number
      required:
        - items
        - total
        - totalAmount
    CreateCostItemRequest:
      type: object
      properties:
        projectId:
          type: integer
        category:
          type: string
          enum: [material, labor, subcontract, expense]
        description:
          type: string
        vendor:
          type: string
          nullable: true
        quantity:
          type: number
          nullable: true
        unit:
          type: string
          nullable: true
        unitPrice:
          type: number
          nullable: true
        amount:
          type: number
        incurredDate:
          type: string
          format: date
        invoiceNumber:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - projectId
        - category
        - description
        - amount
        - incurredDate
    UpdateCostItemRequest:
      type: object
      properties:
        category:
          type: string
          enum: [material, labor, subcontract, expense]
        description:
          type: string
        vendor:
          type: string
          nullable: true
        quantity:
          type: number
          nullable: true
        unit:
          type: string
          nullable: true
        unitPrice:
          type: number
          nullable: true
        amount:
          type: number
        incurredDate:
          type: string
          format: date
        invoiceNumber:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    Budget:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        category:
          type: string
          enum: [material, labor, subcontract, expense]
          description: 予算区分
        description:
          type: string
          description: 内訳名称
        budgetAmount:
          type: number
          description: 予算金額
        actualAmount:
          type: number
          description: 実績金額
        variance:
          type: number
          description: 差異（予算−実績）
        usageRate:
          type: number
          description: 消化率（%）
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - category
        - description
        - budgetAmount
        - actualAmount
        - variance
        - usageRate
        - createdAt
        - updatedAt
    BudgetListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Budget"
        totalBudget:
          type: number
        totalActual:
          type: number
        totalVariance:
          type: number
      required:
        - items
        - totalBudget
        - totalActual
        - totalVariance
    CreateBudgetRequest:
      type: object
      properties:
        projectId:
          type: integer
        category:
          type: string
          enum: [material, labor, subcontract, expense]
        description:
          type: string
        budgetAmount:
          type: number
      required:
        - projectId
        - category
        - description
        - budgetAmount
    UpdateBudgetRequest:
      type: object
      properties:
        category:
          type: string
          enum: [material, labor, subcontract, expense]
        description:
          type: string
        budgetAmount:
          type: number
...
    CostByCategoryResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              label:
                type: string
              amount:
                type: number
              percentage:
                type: number
            required:
              - category
              - label
              - amount
              - percentage
      required:
        - categories
    MonthlyCostsResponse:
      type: object
      properties:
        months:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
                description: YYYY-MM形式
              material:
                type: number
              labor:
                type: number
              subcontract:
                type: number
              expense:
                type: number
              total:
                type: number
            required:
              - month
              - material
              - labor
              - subcontract
              - expense
              - total
      required:
        - months
    BudgetVsActualResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              label:
                type: string
              budget:
                type: number
              actual:
                type: number
              variance:
                type: number
              usageRate:
                type: number
            required:
              - category
              - label
              - budget
              - actual
              - variance
              - usageRate
      required:
        - items