# lib/api-zod/src/generated/types/createCostItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 建設業向け原価管理クラウドシステム API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCostItemRequestCategory } from "./createCostItemRequestCategory";

export interface CreateCostItemRequest {
  projectId: number;
  category: CreateCostItemRequestCategory;
  description: string;
  vendor?: string | null;
  quantity?: number | null;
  unit?: string | null;
  unitPrice?: number | null;
  amount: number;
  incurredDate: Date;
  invoiceNumber?: string | null;
  notes?: string | null;
}