# lib/api-zod/src/generated/types/updateBudgetRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 建設業向け原価管理クラウドシステム API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBudgetRequestCategory } from "./updateBudgetRequestCategory";

export interface UpdateBudgetRequest {
  category?: UpdateBudgetRequestCategory;
  description?: string;
  budgetAmount?: number;
}