# lib/api-zod/src/generated/types/updateCostItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 建設業向け原価管理クラウドシステム API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCostItemRequestCategory } from "./updateCostItemRequestCategory";

export interface UpdateCostItemRequest {
  category?: UpdateCostItemRequestCategory;
  description?: string;
  vendor?: string | null;
  quantity?: number | null;
  unit?: string | null;
  unitPrice?: number | null;
  amount?: number;
  incurredDate?: Date;
  invoiceNumber?: string | null;
  notes?: string | null;
}